requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.10
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-filter==23.5
django-model-utils==4.4.0
django-notifications-hq==1.8.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
environs==10.3.0
et-xmlfile==1.1.0
filelock==3.13.1
flake8==7.0.0
gunicorn==21.2.0
identify==2.5.35
idna==3.6
inflection==0.5.1
isort==5.13.2
jsonfield==3.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.20.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swapper==1.3.0
tblib==3.0.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
